requirements.txt
PyPDF2==3.0.1
langchain==0.0.200
openai==0.27.8
python-dotenv==1.0.0